Refactor tests directory to match zeroae.goblet structure
