+++ tests/chalice/conftest.py
@@ -1,18 +1,5 @@
 import pytest
 from chalice import Chalice
-from environs import Env
-
-
-@pytest.fixture
-def env() -> Env:
-    return Env()
-
-
-@pytest.fixture
-def default_config(env):
-    from zeroae.goblet.chalice import configure
-
-    configure(env)
 
 
 @pytest.fixture
