+++ tests/conftest.py
@@ -1,0 +1,14 @@
+import pytest
+from environs import Env
+
+
+@pytest.fixture
+def env() -> Env:
+    return Env()
+
+
+@pytest.fixture
+def default_config(env):
+    from zeroae.goblet.chalice import configure
+
+    configure(env)
